Feature/nika all (#11)
* Added some dependencies. Added application properties about database. Added liquibase yaml scripts. (#4)

* create entities and repositories

* create dto classes and service methods without realization

---------

+++ service/src/main/java/com/itdev/innovativeproject/dao/entity/DisplayType.java
@@ -1,15 +1,17 @@
 package com.itdev.innovativeproject.dao.entity;
+
 import lombok.Data;
+
 import javax.persistence.*;
 
 @Entity
 @Table(name = "Display_type")
 @Data
-public class DisplayType implements BaseEntity<Integer> {
+public class DisplayType implements BaseEntity<Long> {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
 }
-
